backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
networkx
shapely
google-genai
numpy
python-dotenv
